requirements.txt
-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
crispy-bootstrap5==0.6
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-admin-honeypot==1.1.0
django-allauth==0.44.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.13.0
django-embed-video==1.4.0
django-etc==1.3.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-markdownx==3.0.1
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
facenet-pytorch==2.5.2
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
numpy==1.21.4
oauthlib==3.1.0
openpyxl==3.0.9
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
ranking==0.3.2
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
torch==1.8.2+cpu
torchvision==0.9.2+cpu
typing_extensions==4.0.0
tzlocal==2.1
urllib3==1.26.4
whitenoise==5.2.0
